Connect to GraphQL API
Using Gatsby's `gatsby-source-graqphl` plugin, connect to the in development API(https://github.com/cfpb/wagtail-regulations). Currently this queries for a list of regulations, generates a link on the home page, and generates and individual page for each regulation.

The GraphQL API endpoint can be set in a .env file, using the API_URI variable. This commit introduces an .env.example to use as an exemplar.

gatsby-node queries the API and creates a new static page for each regulation in the database. This will need to be updated to create a subpage for each part on the regulation.